pyproject.toml
[tool.poetry]
name = "py-fast-grpc"
version = "0.3.5"
description = "gRPC simple and easy to use Python framework"
authors = ["Oleg Yurchik"]
license = "MIT License"
readme = "README.md"
packages = [
  { include = "fast_grpc" },
]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.59.3"
grpcio-reflection = "^1.59.3"
grpcio-tools = "^1.59.3"
jinja2 = "^3.1.2"
pydantic = {version = "^2.5.2", extras = ["email"]}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
faker = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
